Refactor execute_batch to execute_query.
Also takes out create_connector and import_module_by_path.

The original create_connector and execute_batch have been stubbed to issue warning, but they will not break external users.
