+++ aperturedb/ParallelQuery.py
@@ -1,92 +1,17 @@
 from __future__ import annotations
-from typing import Callable, List, Tuple, Optional
+from typing import List, Tuple
 from aperturedb import Parallelizer
 import numpy as np
-import json
 import logging
-import math
 import inspect
 
 
 from aperturedb.DaskManager import DaskManager
 from aperturedb.Connector import Connector
 from aperturedb.types import Commands, Blobs, CommandResponses
+from aperturedb.CommonLibrary import execute_query as execute_batch
 
 logger = logging.getLogger(__name__)
-
-
-def execute_batch(q: Commands, blobs: Blobs, db: Connector,
-                  success_statuses: list[int] = [0],
-                  response_handler: Optional[Callable] = None, commands_per_query: int = 1, blobs_per_query: int = 0,
-                  strict_response_validation: bool = False, cmd_index=None) -> Tuple[int, CommandResponses, Blobs]:
-    """
-    Execute a batch of queries, doing useful logging around it.
-    Calls the response handler if provided.
-    This should be used (without the parallel machinery) instead of db.query to keep the response handling consistent, better logging, etc.
-
-    Args:
-        q (Commands): List of commands to execute.
-        blobs (Blobs): List of blobs to send.
-        db (Connector): The database connector.
-        success_statuses (list[int], optional): The list of success statuses. Defaults to [0].
-        response_handler (Callable, optional): The response handler. Defaults to None.
-        commands_per_query (int, optional): The number of commands per query. Defaults to 1.
-        blobs_per_query (int, optional): The number of blobs per query. Defaults to 0.
-        strict_response_validation (bool, optional): Whether to strictly validate the response. Defaults to False.
-
-    Returns:
-        int: The result code.
-            - 0 : if all commands succeeded
-            - 1 : if there was -1 in the response
-            - 2 : For any other code.
-        CommandResponses: The response.
-        Blobs: The blobs.
-    """
-    result = 0
-    logger.debug(f"Query={q}")
-    r, b = db.query(q, blobs)
-    logger.debug(f"Response={r}")
-
-    if db.last_query_ok():
-        if response_handler is not None:
-            try:
-                ParallelQuery.map_response_to_handler(response_handler,
-                                                      q, blobs, r, b, commands_per_query, blobs_per_query,
-                                                      cmd_index)
-            except BaseException as e:
-                logger.exception(e)
-                if strict_response_validation:
-                    raise e
-    else:
-        # Transaction failed entirely.
-        logger.error(f"Failed query = {q} with response = {r}")
-        result = 1
-
-    statuses = {}
-    if isinstance(r, dict):
-        statuses[r['status']] = [r]
-    elif isinstance(r, list):
-        # add each result to a list of the responses, keyed by the response
-        # code.
-        [statuses.setdefault(result[cmd]['status'], []).append(result)
-         for result in r for cmd in result]
-    else:
-        logger.error("Response in unexpected format")
-        result = 1
-
-    # last_query_ok means result status >= 0
-    if result != 1:
-        warn_list = []
-        for status, results in statuses.items():
-            if status not in success_statuses:
-                for wr in results:
-                    warn_list.append(wr)
-        if len(warn_list) != 0:
-            logger.warning(
-                f"Partial errors:\r\n{json.dumps(q)}\r\n{json.dumps(warn_list)}")
-            result = 2
-
-    return result, r, b
 
 
 class ParallelQuery(Parallelizer.Parallelizer):
@@ -111,40 +36,6 @@ class ParallelQuery(Parallelizer.Parallelizer):
     @classmethod
     def getSuccessStatus(cls):
         return cls.success_statuses
-
-    @classmethod
-    def map_response_to_handler(cls, handler, query, query_blobs,  response, response_blobs,
-                                commands_per_query, blobs_per_query, cmd_index_offset):
-        # We could potentially always call this handler function
-        # and let the user deal with the error cases.
-        blobs_returned = 0
-        for i in range(math.ceil(len(query) / commands_per_query)):
-            start = i * commands_per_query
-            end = start + commands_per_query
-            blobs_start = i * blobs_per_query
-            blobs_end = blobs_start + blobs_per_query
-
-            b_count = 0
-            if issubclass(type(response), list):
-                for req, resp in zip(query[start:end], response[start:end]):
-                    for k in req:
-                        blob_returning_commands = ["FindImage", "FindBlob", "FindVideo",
-                                                   "FindDescriptor", "FindBoundingBox"]
-                        if k in blob_returning_commands and "blobs" in req[k] and req[k]["blobs"]:
-                            count = resp[k]["returned"]
-                            b_count += count
-
-            # The returned blobs need to be sliced to match the
-            # returned entities per command in query.
-            handler(
-                query[start:end],
-                query_blobs[blobs_start:blobs_end],
-                response[start:end] if issubclass(
-                    type(response), list) else response,
-                response_blobs[blobs_returned:blobs_returned + b_count] if
-                len(response_blobs) >= blobs_returned + b_count else None,
-                None if cmd_index_offset is None else cmd_index_offset + i)
-            blobs_returned += b_count
 
     def __init__(self, db: Connector, dry_run: bool = False):
         super().__init__()
@@ -463,3 +354,10 @@ class ParallelQuery(Parallelizer.Parallelizer):
             print("Sample of data to be ingested:")
             for i in range(sample_count):
                 print(self.generator[i])
+
+
+def execute_batch(*args, **kwargs):
+    from aperturedb.CommonLibrary import execute_query, issue_deprecation_warning
+    issue_deprecation_warning(
+        "ParallelQuery.execute_batch", "CommonLibrary.execute_query")
+    return execute_query(*args, **kwargs)
