+++ examples/dask/ingest_dask.py
@@ -19,7 +19,7 @@ if __name__ == '__main__':
         FILENAME, blocksize=os.path.getsize(FILENAME) // numthreads)
 
     def process(df):
-        from aperturedb.Utils import create_connector
+        from aperturedb.CommonLibrary import create_connector
         from aperturedb.ParallelLoader import ParallelLoader
         db = create_connector()
         loader = ParallelLoader(db)
