+++ examples/dask/ingest_loader.py
@@ -1,7 +1,7 @@
 import os
 from aperturedb.EntityDataCSV import EntityDataCSV
 from aperturedb.ParallelLoader import ParallelLoader
-from aperturedb.Utils import create_connector
+from aperturedb.CommonLibrary import create_connector
 import typer
